+++ samples/sample.cpp
@@ -15,17 +15,33 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "TaskScheduler.h"
+
+class SampleTask : public enki::ITaskSet
+{
+public:
+	SampleTask() = default;
+
+	void ExecuteRange( enki::TaskSetPartition range, uint32_t threadIndex ) override
+	{
+		m_task( range.start, range.end, threadIndex, m_taskContext );
+	}
+
+	b2TaskCallback* m_task = nullptr;
+	void* m_taskContext = nullptr;
+};
+
 static void* EnqueueTask( b2TaskCallback* task, int32_t itemCount, int32_t minRange, void* taskContext, void* userContext )
 {
 	Sample* sample = static_cast<Sample*>( userContext );
-	if ( sample->m_taskCount < maxTasks )
+	if ( sample->m_taskCount < Sample::m_maxTasks )
 	{
 		SampleTask& sampleTask = sample->m_tasks[sample->m_taskCount];
 		sampleTask.m_SetSize = itemCount;
 		sampleTask.m_MinRange = minRange;
 		sampleTask.m_task = task;
 		sampleTask.m_taskContext = taskContext;
-		sample->m_scheduler.AddTaskSetToPipe( &sampleTask );
+		sample->m_scheduler->AddTaskSetToPipe( &sampleTask );
 		++sample->m_taskCount;
 		return &sampleTask;
 	}
@@ -44,7 +60,7 @@ static void FinishTask( void* taskPtr, void* userContext )
 	{
 		SampleTask* sampleTask = static_cast<SampleTask*>( taskPtr );
 		Sample* sample = static_cast<Sample*>( userContext );
-		sample->m_scheduler.WaitforTask( sampleTask );
+		sample->m_scheduler->WaitforTask( sampleTask );
 	}
 }
 
@@ -78,7 +94,10 @@ static void TestMathCpp()
 
 Sample::Sample( Settings& settings )
 {
-	m_scheduler.Initialize( settings.workerCount );
+	m_scheduler = new enki::TaskScheduler;
+	m_scheduler->Initialize( settings.workerCount );
+
+	m_tasks = new SampleTask[m_maxTasks];
 	m_taskCount = 0;
 
 	m_threadCount = 1 + settings.workerCount;
@@ -111,6 +130,9 @@ Sample::~Sample()
 {
 	// By deleting the world, we delete the bomb, mouse joint, etc.
 	b2DestroyWorld( m_worldId );
+
+	delete m_scheduler;
+	delete[] m_tasks;
 }
 
 void Sample::DrawTitle( const char* string )
@@ -242,7 +264,7 @@ void Sample::Step( Settings& settings )
 			timeStep = 0.0f;
 		}
 
-		if (g_draw.m_showUI)
+		if ( g_draw.m_showUI )
 		{
 			g_draw.DrawString( 5, m_textLine, "****PAUSED****" );
 			m_textLine += m_textIncrement;
